timer: cadence_ttc: Implement increased bit widths
Allowing for up to 32b support.

+++ include/hw/timer/cadence_ttc.h
@@ -29,9 +29,9 @@ typedef struct {
 
     uint32_t reg_clock;
     uint32_t reg_count;
-    uint32_t reg_value;
-    uint16_t reg_interval;
-    uint16_t reg_match[3];
+    uint64_t reg_value;
+    uint32_t reg_interval;
+    uint32_t reg_match[3];
     uint32_t reg_intr;
     uint32_t reg_intr_en;
     uint32_t reg_event_ctrl;
